Woo: Update local state upon product create
This adds code that is compatible with action-list, etc, by extending
the product updated action with a completion action that can carry on a
success action that was given to the request. This ensures that every
time a product is successfully created, that it will be updated in the
local state.